hotel_service/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
mysqlclient
python-dotenv
drf-spectacular